Split auxiliary further
