+++ nrpy/infrastructures/BHaH/nrpyelliptic/__init__.py
@@ -1,1 +1,1 @@
-from . import auxiliary, diagnostics, initial_data, rhs_eval
+from . import constant_source_terms_to_auxevol, diagnostics, initial_data, rhs_eval, variable_wavespeed_gfs
